requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.39.0
accelerate>=0.28.0
peft>=0.10.0
pandas>=2.0.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.66.0
evaluate>=0.4.1
rouge-score>=0.1.2
sacrebleu>=2.4.0
